feat: same thing, different structure
