--- packages/frontend/src/presentation/components/button/Button.css
@@ -1,18 +1,0 @@
-.button {
-  display:block;
-  float:left;
-  margin:0 7px 0 0;
-  background-color:#f5f5f5;
-  border:1px solid #dedede;
-  border-top:1px solid #eee;
-  border-left:1px solid #eee;
-  font-size:100%;
-  line-height:130%;
-  text-decoration:none;
-  font-weight:bold;
-  color:#565656;
-  cursor:pointer;
-  width:auto;
-  overflow:visible;
-  padding:4px 10px 3px 7px;
-}
