+++ packages/frontend/src/presentation/components/button/Button.tsx
@@ -1,6 +1,24 @@
 import React from 'react';
-import { View } from 'react-native';
+import { StyleSheet, Text, TouchableOpacity } from 'react-native';
 
-export const Button: React.FC = () => {
-  return <View />;
+interface ButtonProps {
+  onPress: () => void;
+  label: string;
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+    height: 20,
+  },
+});
+
+export const Button: React.FC<ButtonProps> = (props: ButtonProps) => {
+  return (
+    <TouchableOpacity style={styles.container} onPress={props.onPress}>
+      <Text>{props.label}</Text>
+    </TouchableOpacity>
+  );
 };
