--- packages/frontend/src/presentation/components/button/Button.web.tsx
@@ -1,8 +1,0 @@
-import React, { ButtonHTMLAttributes } from 'react';
-import './Button.css';
-
-export interface ButtonProps extends ButtonHTMLAttributes<HTMLButtonElement> {}
-
-export const Button: React.FC<ButtonProps> = (props: ButtonProps) => {
-  return <button {...props} className={`button ${props.className}`.trim()} />;
-};
