--- packages/frontend/src/presentation/screens/CardSearch/CardSearchResults.css
@@ -1,3 +1,0 @@
-ul li {
-  cursor: pointer;
-}
